requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
alabaster==0.7.12
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
cchardet==2.1.7
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.7
Django==3.1.7
django-db-connection-pool==1.0.7
djangorestframework==3.12.2
dnspython==2.1.0
docutils==0.16
exrex==0.10.5
fake-useragent==0.1.11
filelock==3.0.12
fire==0.4.0
future==0.18.2
geoip2==4.1.0
greenlet==1.0.0
idna==2.10
imagesize==1.2.0
IPy==1.1
Jinja2==2.11.3
kombu==5.0.2
loguru==0.5.3
lxml==4.6.3
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.1.0
netaddr==0.7.19
numpy==1.20.2
packaging==20.9
pandas==1.2.3
prompt-toolkit==3.0.18
psutil==5.8.0
pycparser==2.20
pycurl==7.43.0.2
pyee==8.1.0
Pygments==2.9.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyppeteer==0.2.5
pyShodan==0.2.3
PySocks==1.7.1
python-dateutil==2.8.1
python-magic==0.4.22
python3-nmap==1.5.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
shodan==1.25.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.5
sqlparse==0.4.1
supervisor==4.2.2
tenacity==7.0.0
termcolor==1.1.0
tld==0.12.5
tldextract==3.1.0
tqdm==4.59.0
treelib==1.6.1
typing-extensions==3.10.0.0
urllib3==1.25.11
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
XlsxWriter==1.4.0
xlwt==1.3.0
yarl==1.6.3
